move result definition to zarrextra, with some documentation
