+++ docs/docs/core/error-handling.mdx
@@ -14,6 +14,12 @@ This aspect of the design may be subject to review, and community feedback is en
 
 :::
 
+:::note Result implementation
+
+The `Result` type is implemented in `@spatialdata/zarrextra` and re-exported from `@spatialdata/core` for convenience. This is currently a custom implementation for simplicity and to avoid dependencies. We may review using an existing Result library (such as `neverthrow` or `ts-result`) in the future, but for now this provides a lightweight, dependency-free solution.
+
+:::
+
 ## The Result Type
 
 ```ts
