+++ packages/core/src/types.ts
@@ -12,46 +12,6 @@
  * 4. Consider whether a type belongs here vs. in a more specific module
  */
 
-// ============================================
-// Result Type
-// ============================================
-
-/**
- * A Result type for explicit error handling without exceptions.
- * Inspired by Rust's Result<T, E>.
- */
-export type Result<T, E = Error> = 
-  | { ok: true; value: T }
-  | { ok: false; error: E };
-
-/** Create a successful Result */
-export const Ok = <T>(value: T): Result<T, never> => ({ ok: true, value });
-
-/** Create a failed Result */
-export const Err = <E>(error: E): Result<never, E> => ({ ok: false, error });
-
-/** Type guard for Ok results */
-export const isOk = <T, E>(result: Result<T, E>): result is { ok: true; value: T } => result.ok;
-
-/** Type guard for Err results */
-export const isErr = <T, E>(result: Result<T, E>): result is { ok: false; error: E } => !result.ok;
-
-/**
- * Unwrap a Result, throwing if it's an error.
- * Use when you want to convert back to exception-based error handling.
- */
-export const unwrap = <T, E>(result: Result<T, E>): T => {
-  if (result.ok) return result.value;
-  throw result.error instanceof Error ? result.error : new Error(String(result.error));
-};
-
-/**
- * Unwrap a Result with a default value for errors.
- */
-export const unwrapOr = <T, E>(result: Result<T, E>, defaultValue: T): T => {
-  return result.ok ? result.value : defaultValue;
-};
-
 import type * as ad from 'anndata.js';
 import type * as zarr from 'zarrita';
 import type { ZarrTree } from '@spatialdata/zarrextra';
@@ -115,6 +75,11 @@ export type ZGroup = zarr.Group<Store>;
 // These are used in SDataProps and models, so we keep them accessible from core/types
 export type { ZarrTree, LazyZarrArray, ZAttrsAny } from '@spatialdata/zarrextra';
 export { ATTRS_KEY, ZARRAY_KEY } from '@spatialdata/zarrextra';
+
+// Re-export Result type and utilities from zarrextra for convenience
+// Result is used throughout core for explicit error handling
+export type { Result } from '@spatialdata/zarrextra';
+export { Ok, Err, isOk, isErr, unwrap, unwrapOr } from '@spatialdata/zarrextra';
 
 
 /**
